+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to
 
 - 📝Contributing.md #352
 - 🌐(frontend) add localization to editor #268
+- ✨Public and restricted doc editable #357
 
 ## Fixed